% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfsrouter-package.R
\docType{data}
\name{berlin_gtfs}
\alias{berlin_gtfs}
\title{berlin_gtfs}
\format{A list of five \pkg{data.table} items necessary for routing:
\itemize{
\item routes
\item trips
\item stop_times
\item stops
\item transfers
}}
\description{
Sample GTFS data from Verkehrsverbund Berlin-Brandenburg street, reduced to U
and S Bahn only (underground and overground trains), and between the hours of
12:00-13:00. Only those components of the GTFS data necessary for routing
have been retained.
}
\note{
Can be re-created with the script in
\url{https://github.com/ATFutures/gtfs-router/blob/master/data-raw/data-script.Rmd}.
}
\keyword{datasets}
