% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies_to_stop_times.R
\name{frequencies_to_stop_times}
\alias{frequencies_to_stop_times}
\title{frequencies_to_stop_times}
\usage{
frequencies_to_stop_times(gtfs)
}
\arguments{
\item{gtfs}{A set of GTFS data returned from \link{extract_gtfs}.}
}
\value{
The input GTFS data with data from the 'frequencies' table converted
to equivalent 'arrival_time' and 'departure_time' values in \code{stop_times}.
}
\description{
Convert a GTFS 'frequencies' table to equivalent 'stop_times' that can be
used for routing.
}
\examples{
\dontrun{
# Presume an input feed has been created and includes a "frequencies" table:
gtfs2 <- frequencies_to_stop_times (gtfs)
# "gtfs2" will then have an expanded "stop_times" table, with all
# "frequencies" entries converted to equivalent absolute stop times.
}
}
\seealso{
Other augment: 
\code{\link{gtfs_transfer_table}()}
}
\concept{augment}
