% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrone.R
\name{gtfs_isochrone}
\alias{gtfs_isochrone}
\title{gtfs_isochrone}
\usage{
gtfs_isochrone(gtfs, from, start_time, end_time, day = NULL,
  route_pattern = NULL, hull_alpha = 0.1, quiet = FALSE)
}
\arguments{
\item{gtfs}{A set of GTFS data returned from \link{extract_gtfs} or, for more
efficient queries, pre-processed with \link{gtfs_timetable}.}

\item{from}{Name of start station}

\item{start_time}{Desired departure time at \code{from} station, either in seconds
after midnight, a vector of two or three integers (hours, minutes) or (hours,
minutes, seconds), an object of class \link{difftime}, \pkg{hms}, or
\pkg{lubridate}.}

\item{end_time}{End time to calculate isochrone}

\item{day}{Day of the week on which to calculate route, either as an
unambiguous string (so "tu" and "th" for Tuesday and Thursday), or a number
between 1 = Sunday and 7 = Saturday. If not given, the current day will be
used. (Not used if \code{gtfs} has already been prepared with
\link{gtfs_timetable}.)}

\item{route_pattern}{Using only those routes matching given pattern, for
example, "^U" for routes starting with "U" (as commonly used for underground
or subway routes. (Parameter not used at all if \code{gtfs} has already been
prepared with \link{gtfs_timetable}.)}

\item{hull_alpha}{alpha value of non-convex hulls returned as part of result
(see ?alphashape::ashape for details).}

\item{quiet}{Set to \code{TRUE} to suppress screen messages (currently just
regarding timetable construction).}
}
\value{
square matrix of distances between nodes
}
\description{
Calculate a single isochrone from a given start station, returning the list
of all stations reachable to the specified \code{end_time}.
}
