% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrones.R
\name{gtfs_isochrones}
\alias{gtfs_isochrones}
\title{gtfs_isochrones}
\usage{
gtfs_isochrones(gtfs, from, time_incr, quiet = FALSE)
}
\arguments{
\item{gtfs}{A set of GTFS data returned from \link{extract_gtfs} or, for more
efficient queries, pre-processed with \link{gtfs_timetable}.}

\item{from}{Name of start station}

\item{time_incr}{Temporal increment of isochrone in seconds.}

\item{quiet}{If \code{FALSE}, display progress bar and other information.}
}
\value{
don't know yet
}
\description{
Calculate all possible isochrones from a given start station.
}
\examples{
berlin_gtfs_to_zip ()
f <- file.path (tempdir (), "vbb.zip")
g <- extract_gtfs (f)
g <- gtfs_timetable (g)
from <- "Schonlein"
all_routes <- gtfs_isochrones (g, from, time_incr = 600)
}
