% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable.R
\name{gtfs_median_graph}
\alias{gtfs_median_graph}
\title{gtfs_median_graph}
\usage{
gtfs_median_graph(timetable, gtfs)
}
\arguments{
\item{timetable}{Median timetable returned from \link{gtfs_median_timetable}.}

\item{gtfs}{GTFS data set with timetable obtained from \link{extract_gtfs}
and \link{gtfs_timetable}.}
}
\value{
A \code{data.frame} containing 3 columns: the departure and arrival
stations as integer indices into \code{gtfs$stop_ids}, and a column of edge
lengths measured as median durations between each departure and arrival
station.
}
\description{
Construct graph from median timetable including all transfers
}
\examples{
berlin_gtfs_to_zip ()
f <- file.path (tempdir (), "vbb.zip")
gtfs <- extract_gtfs (f)
gtfs <- gtfs_timetable (gtfs, day = 1:7)
tt <- gtfs_median_timetable (gtfs)
graph <- gtfs_median_graph (tt, gtfs)
d <- gtfs_median_isochrones (graph, station = 100)
}
