% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrones.R
\name{gtfs_median_isochrones}
\alias{gtfs_median_isochrones}
\title{gtfs_median_isochrones}
\usage{
gtfs_median_isochrones(graph, station)
}
\arguments{
\item{graph}{A GTFS median graph, processed from a GTFS data set with
\link{extract_gtfs}, \link{gtfs_timetable}, \link{gtfs_median_timetable}, and
\link{gtfs_median_graph}}

\item{station}{Integer defining start station as index into \code{gtfs$stop_ids}}
}
\value{
Vector of median times taken to reach each station in
\code{gtfs$stop_ids}.
}
\description{
Calculate all possible isochrones from a given start station using the median
graph of a GTFS data set.
}
\examples{
berlin_gtfs_to_zip ()
f <- file.path (tempdir (), "vbb.zip")
gtfs <- extract_gtfs (f)
gtfs <- gtfs_timetable (gtfs, day = 1:7)
tt <- gtfs_median_timetable (gtfs)
graph <- gtfs_median_graph (tt, gtfs)
d <- gtfs_median_isochrones (graph, station = 100)
}
