% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable.R
\name{gtfs_median_timetable}
\alias{gtfs_median_timetable}
\title{gtfs_median_timetable}
\usage{
gtfs_median_timetable(gtfs)
}
\arguments{
\item{gtfs}{A set of GTFS data with timetable returned from
\link{extract_gtfs} and \link{gtfs_timetable}.}
}
\value{
A \code{data.frame} containing the median timetable.
}
\description{
Convert GTFS data into format able to be used to calculate routes.
}
\examples{
berlin_gtfs_to_zip ()
f <- file.path (tempdir (), "vbb.zip")
gtfs <- extract_gtfs (f)
gtfs <- gtfs_timetable (gtfs, day = 1:7)
tt <- gtfs_median_timetable (gtfs)
graph <- gtfs_median_graph (tt, gtfs)
d <- gtfs_median_isochrones (graph, station = 100)
}
