% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route-stops.R
\name{gtfs_route_stops}
\alias{gtfs_route_stops}
\title{gtfs_route_stops}
\usage{
gtfs_route_stops(gtfs, route_id)
}
\arguments{
\item{gtfs}{A set of GTFS data returned from \link{extract_gtfs} or, for more
efficient queries, pre-processed with \link{gtfs_timetable}.}

\item{route_id}{Character vector of \code{route_id} from the GTFS "routes.txt" table}
}
\value{
List of \code{data.frame} objects containing sequences of \code{stop_id} and
\code{stop_name} values for that route.
}
\description{
Get lists of stop sequences for a specified route
}
\examples{
berlin_gtfs_to_zip ()
f <- file.path (tempdir (), "vbb.zip")
g <- extract_gtfs (f)
route_id <- "10141_109"
r <- gtfs_route_stops (g, route_id)

}
