% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable.R
\name{gtfs_timetable}
\alias{gtfs_timetable}
\title{gtfs_timetable}
\usage{
gtfs_timetable(gtfs, day = NULL)
}
\arguments{
\item{gtfs}{A set of GTFS data returned from \link{extract_gtfs}.}

\item{day}{Day of the week on which to calculate route, either as an
unambiguous string (so "tu" and "th" for Tuesday and Thursday), or a number
between 1 = Sunday and 7 = Saturday. If not given, the current day will be
used.}
}
\value{
The input data with an addition items, \code{timetable}, \code{stations}, and
\code{trips}, containing data formatted for more efficient use with
\link{gtfs_route} (see Note).
}
\description{
Convert GTFS data into format able to be used to calculate routes.
}
\note{
This function is merely provided to speed up calls to the primary
function, \link{gtfs_route}. If the input data to that function do not
include a formatted \code{timetable}, it will be calculated anyway, but queries in
that case will generally take longer.
}
