% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfsrouter-package.R
\docType{package}
\name{gtfsrouter}
\alias{gtfsrouter}
\alias{gtfsrouter-package}
\title{gtfsrouter}
\description{
Routing engine for GTFS (General Transit Feed Specification) data, including
one-to-one and one-to-many routing routines.
}
\section{Main Functions}{

\itemize{
\item \code{\link[=gtfs_route]{gtfs_route()}}: Route between given start and end stations using a
nominated GTFS data set.
\item \code{\link[=go_home]{go_home()}}: Automatic routing between work and home stations specified
with local environmental variables
\item \code{\link[=go_to_work]{go_to_work()}}: Automatic routing between work and home stations
specified with local environmental variables
\item \code{\link[=gtfs_traveltimes]{gtfs_traveltimes()}}: One-to-many routing from a nominated start
station to all stations reachable within a specified travel duration.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UrbanAnalyst/gtfsrouter}
  \item Report bugs at \url{https://github.com/UrbanAnalyst/gtfsrouter/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

Authors:
\itemize{
  \item Marcin Stepniak \email{marcinstepniak@ucm.es} (\href{https://orcid.org/0000-0002-6489-5443}{ORCID})
}

Other contributors:
\itemize{
  \item Alexandra Kapp \email{alk@mobilityinstitute.com} [contributor]
}

}
\concept{package}
