% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{koeln_gtfs_to_zip}
\alias{koeln_gtfs_to_zip}
\title{koeln_gtfs_to_zip}
\usage{
koeln_gtfs_to_zip()
}
\value{
Nothing
}
\description{
Write a zip archive of the internal package data, \link{koeln_gtfs} to
a file named "vrs.zip" to \code{tempdir()}.
}
