% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.gtfs_ischrone}
\alias{plot.gtfs_ischrone}
\alias{plot.gtfs_isochrone}
\title{plot.gtfs_isochrone}
\usage{
\method{plot}{gtfs_isochrone}(x, ...)
}
\arguments{
\item{x}{object to be plotted}

\item{...}{ignored here}
}
\description{
plot.gtfs_isochrone
}
\seealso{
Other additional: 
\code{\link{go_home}()},
\code{\link{go_to_work}()},
\code{\link{process_gtfs_local}()},
\code{\link{summary.gtfs}()}
}
\concept{additional}
