% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.gtfs}
\alias{summary.gtfs}
\title{summary.gtfs}
\usage{
\method{summary}{gtfs}(object, ...)
}
\arguments{
\item{object}{A \code{gtfs} object to be summarised}

\item{...}{ignored here}
}
\description{
summary.gtfs
}
\examples{
berlin_gtfs_to_zip ()
f <- file.path (tempdir (), "vbb.zip")
g <- extract_gtfs (f)
summary (g)
g <- gtfs_timetable (g)
summary (g) # also summarizes additional timetable information
}
