% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{berlin_gtfs_to_zip}
\alias{berlin_gtfs_to_zip}
\title{berlin_gtfs_to_zip}
\usage{
berlin_gtfs_to_zip()
}
\value{
Nothing
}
\description{
Write a zip archive of the internal package data, \link{berlin_gtfs} to
a file named "vbb.zip" to \code{tempdir()}.
}
