% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-gtfs.R
\name{extract_gtfs}
\alias{extract_gtfs}
\title{extract_gtfs}
\usage{
extract_gtfs(filename = NULL, quiet = FALSE)
}
\arguments{
\item{filename}{Name of GTFS archive}

\item{quiet}{If \code{FALSE}, display progress information on screen}
}
\value{
List of several \pkg{data.table} objects corresponding to the tables
present in the nominated GTFS data set.
}
\description{
Extract data from a GTFS \code{zip} archive.
}
\examples{
berlin_gtfs_to_zip () # Write sample feed from Berlin, Germany to tempdir
f <- file.path (tempdir (), "vbb.zip") # name of feed
gtfs <- extract_gtfs (f)

}
