% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{extract_gtfs}
\alias{extract_gtfs}
\title{extract_gtfs}
\usage{
extract_gtfs(filename = NULL)
}
\arguments{
\item{filename}{Name of GTFS archive}
}
\value{
List of 2 \pkg{data.table} objects, one for "stop_times" and one for
"transfers"
}
\description{
Extract "stop_times" and "transfers" table from a GTFS \code{zip} archive.
}
