% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies_to_stop_times.R
\name{frequencies_to_stop_times}
\alias{frequencies_to_stop_times}
\title{frequencies_to_stop_times}
\usage{
frequencies_to_stop_times(gtfs)
}
\arguments{
\item{gtfs}{A set of GTFS data returned from \link{extract_gtfs}.}
}
\value{
The input gtfs data with data from \code{frequencies} replaced by
arrival_time and departure_time in \code{stop_times}
}
\description{
Convert gtfs frequencies to stop_times
}
