% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{gtfs_route}
\alias{gtfs_route}
\title{gtfs_route}
\usage{
gtfs_route(graph, from, to)
}
\arguments{
\item{graph}{\code{data.frame} or equivalent object representing the network
graph (see Details)}

\item{from}{Vector or matrix of points \strong{from} which route distances are to
be calculated (see Details)}

\item{to}{Vector or matrix of points \strong{to} which route distances are to be
calculated (see Details)}
}
\value{
square matrix of distances between nodes
}
\description{
Calculate matrix of pair-wise distances between points.
}
