% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfsrouter-package.R
\docType{package}
\name{gtfsrouter}
\alias{gtfsrouter}
\alias{gtfsrouter-package}
\title{gtfsrouter}
\description{
Find the quickest route between two points at a particular start time with
GTFS feed.
}
