% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfsrouter-package.R
\docType{package}
\name{gtfsrouter}
\alias{gtfsrouter}
\title{gtfsrouter}
\description{
Routing engine for GTFS (General Transit Feed Specification) data, including
one-to-one and one-to-many routing routines.
}
\section{Main Functions}{

\itemize{
\item \code{\link[=gtfs_route]{gtfs_route()}}: Route between given start and end stations using a
nominated GTFS data set.
\item \code{\link[=go_home]{go_home()}}: Automatic routing between work and home stations specified
with local environmental variables
\item \code{\link[=go_to_work]{go_to_work()}}: Automatic routing between work and home stations
specified with local environmental variables
\item \code{\link[=gtfs_isochrone]{gtfs_isochrone()}}: One-to-many routing from a nominated start station
to all stations reachable within a specified travel duration.
}
}

