% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrone.R
\name{plot.gtfs_ischrone}
\alias{plot.gtfs_ischrone}
\alias{plot.gtfs_isochrone}
\title{plot.gtfs_isochrone}
\usage{
\method{plot}{gtfs_isochrone}(x, ..., hull_alpha = 0.1,
  show_all = FALSE)
}
\arguments{
\item{x}{object to be plotted}

\item{...}{ignored here}

\item{hull_alpha}{alpha value of non-convex hulls (see ?alphashape::ashape
for details).}

\item{show_all}{If \code{TRUE}, all other stations beyond the isochrone are also
plotted (currently not implemented).}
}
\description{
plot.gtfs_isochrone
}
