% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.gtfs_ischrone}
\alias{plot.gtfs_ischrone}
\alias{plot.gtfs_isochrone}
\title{plot.gtfs_isochrone}
\usage{
\method{plot}{gtfs_isochrone}(x, ...)
}
\arguments{
\item{x}{object to be plotted}

\item{...}{ignored here}
}
\description{
plot.gtfs_isochrone
}
